refactor(tab-details): add tab details model
Added tab details model
