+++ models/community/community_tabs_model.ts
@@ -1,11 +1,6 @@
+import { TabDetails } from '@/models/tab_details/tab_details';
+
 export interface CommunityTabsModel {
   name: string;
   details: TabDetails[];
-}
-
-export interface TabDetails {
-  name: string;
-  description: string;
-  images: string[];
-  members: string[];
 }
