+++ models/tab_details/tab_details.ts
@@ -1,0 +1,6 @@
+export interface TabDetails {
+  name: string;
+  description: string;
+  images: string[];
+  members: string[];
+}
